Add anchor for denoting start of changelog content
This allows the parser to seek past the changelog's header (if it exists) and
parse only the changelog content.

In a later commit, the header text on disk will be prepended to the changelog
file before the latest release. This is necessary to ensure that the updated
header text is prepended correctly, especially when the header gets shorter